requirements.txt
torch>=2.0.0
transformers>=4.35.0
accelerate>=0.24.0
einops>=0.7.0

biopython>=1.81
biotite>=0.37.0

numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0
scikit-learn>=1.3.0

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
aiofiles>=23.2.1

h5py>=3.9.0
tqdm>=4.66.0
pydantic>=2.4.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

pytest>=7.4.0
httpx>=0.25.0
